requirements.txt
# Superannuation Advisory Agent - Complete Dependencies
# Production-ready multi-country retirement advisory system

# Core Framework
# Pin to 1.51.0 - older versions (1.37-1.40) have "Bad setIn index" widget state bugs
streamlit==1.51.0

# Databricks & MLflow
mlflow>=2.8.0
databricks-sdk>=0.12.0

# Data Processing
pandas>=2.0.0

# Data Generation (for demos)
faker>=28.0.0

# Visualization (for monitoring dashboards)
plotly>=5.14.0

# Type Checking
pydantic>=2.0.0

# Configuration Management (Phase 1)
PyYAML>=6.0.0

# Template Engine (Phase 2)
Jinja2>=3.1.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
